fix some warnings

+++ src/cron/wiltoncall_cron.cpp
@@ -64,7 +64,7 @@ sl::support::optional<sl::io::span<char>> cron_start(sl::io::span<const char> da
                 // output will be ignored
                 char* out = nullptr;
                 int out_len = 0;
-                auto err = wiltoncall_runscript(engine.c_str(), engine.length(), 
+                auto err = wiltoncall_runscript(engine.c_str(), static_cast<int> (engine.length()), 
                         str->c_str(), static_cast<int> (str->length()),
                         std::addressof(out), std::addressof(out_len));
                 if (nullptr != err) {
